extract kubeconfig, client to apiserver, move the leaderelection as subpackage of apiserver
